pyproject.toml
[project]
name = "collect-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.34.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
]
